pyproject.toml
[project]
name = "ai-research-multi-agent"
version = "1.0.0"
description = "研究開発用AI Agent"
authors = [
    {name = "Yusuke Sakai (GitHub: Yagami360)", email = "y034112@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic~=0.69.0",
    "claude-code-sdk~=0.0.25",
    "google-genai>=1.41.0,<2.0.0",
    "google-adk~=1.17.0",
    "requests>=2.32.4,<3.0.0",
    "pydantic~=2.11.9",
    "pydantic-settings~=2.11.0",
    "pyyaml~=6.0.0",
    "huggingface-hub>=0.35.3",
    "python-dotenv~=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest~=7.0.0",
    "black>=22.0.0,<24.0.0",
    "isort~=5.12.0",
    "flake8~=6.0.0",
    "mypy~=1.0.0",
    "types-PyYAML~=6.0.0",
    "types-requests~=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 150
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 150
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [".venv/", "venv/", "__pycache__/", ".git/", ".mypy_cache/", ".pytest_cache/", ".old/", ".debug/", ".logs/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
